requirements.txt
streamlit
pandas
openpyxl
google-generativeai
deep-translator
googletrans
python-dotenv
supabase